[EPMEDU-1090]: Rename Help page to FAQ (#114)
