+++ feature/tabsflow/moreflow/faq/src/main/java/com/epmedu/animeal/tabs/more/faq/viewmodel/FAQState.kt
@@ -1,0 +1,5 @@
+package com.epmedu.animeal.tabs.more.faq.viewmodel
+
+data class FAQState(
+    val appVersionName: String = ""
+)
