--- feature/tabsflow/moreflow/help/src/main/java/com/epmedu/animeal/tabs/more/help/viewmodel/HelpState.kt
@@ -1,5 +1,0 @@
-package com.epmedu.animeal.tabs.more.help.viewmodel
-
-data class HelpState(
-    val appVersionName: String = ""
-)
