+++ library/common/src/main/java/com/epmedu/animeal/common/route/MoreRoute.kt
@@ -4,7 +4,7 @@ enum class MoreRoute {
     More,
     Profile,
     Donate,
-    Help,
+    FAQ,
     About,
     Account
 }
