+++ library/resources/src/main/res/values/strings.xml
@@ -21,7 +21,7 @@
 
     <string name="page_profile">Profile Page</string>
     <string name="page_donate">Donate</string>
-    <string name="page_help">Help</string>
+    <string name="page_faq">FAQ</string>
     <string name="page_about">About</string>
     <string name="page_account">Account</string>
     <string name="page_about_detailed">About (terms and conditions)</string>
@@ -80,7 +80,7 @@
 
     <string name="willfeed_timeleft_msg">You will have 1 hour to feed the point</string>
 
-    <string name="help_page_app_version">App version: %s</string>
+    <string name="faq_page_app_version">App version: %s</string>
 
     <string name="done">Done</string>
     <string name="ok">OK</string>
