+++ settings.gradle.kts
@@ -61,7 +61,7 @@ include(":feature:tabsflow:analytics")
 include(":feature:tabsflow:moreflow:about")
 include(":feature:tabsflow:moreflow:account")
 include(":feature:tabsflow:moreflow:donate")
-include(":feature:tabsflow:moreflow:help")
+include(":feature:tabsflow:moreflow:faq")
 include(":feature:tabsflow:moreflow:host")
 include(":feature:tabsflow:moreflow:more")
 include(":feature:tabsflow:moreflow:profile")
